Feature/sync all data



See merge request !453

+++ tests/unit_test/c/unittest_modules/meta_mem_cache/unittests/mock_function.c
@@ -117,7 +117,7 @@ int check_and_copy_file(const char *srcpath, const char *tarpath,
 	return MOCK_RETURN_VAL;
 }
 
-int fetch_block_path(char *pathname, ino_t this_inode, long long block_num)
+int fetch_block_path(char *pathname, ino_t this_inode, int64_t block_num)
 {
 	return 0;
 }
